tests: add region bounds test to existing test_myfunc executable; revert core changes; fix CMake to avoid extra test target

+++ myfunc.cc
@@ -77,8 +77,7 @@ struct Region {
   // Note: Region does not own the memory, caller is responsible for cleanup
 
   std::optional<unsigned int> read(unsigned int addr, int size) {
-    // Ensure the entire access fits within the region
-    if (addr < start_ || addr + static_cast<unsigned int>(size) > start_ + size_) {
+    if (addr < start_ || addr + size > start_ + size_) {
       return std::nullopt;
     }
     unsigned int offset = addr - start_;
@@ -90,8 +89,7 @@ struct Region {
   }
 
   bool write(unsigned int addr, int size, unsigned int value) {
-    // Ensure the entire access fits within the region
-    if (addr < start_ || addr + static_cast<unsigned int>(size) > start_ + size_) {
+    if (addr < start_ || addr + size > start_ + size_) {
       return false;
     }
     unsigned int offset = addr - start_;
